Add files via upload
This visual is based on the current INFORM visualization built with the support of the ARC. It is a subset of the INFORM data  that focuses on 8 countries: Armenia, Azerbaijan, Georgia, Kyrgyzstan,Kazakhstan,Tajikistan, turkmenistan, and Uzbekistan.   

The level of granularity of this data is at subnational level, first administrative level.